requirements.txt
fastapi
uvicorn
deep-translator
groq
requests
python-dotenv